+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @clerk/ui
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [[`427fcdeaa`](https://github.com/clerk/javascript/commit/427fcdeaaba4e77273be29b4d7cca43f9aa18693)]:
+  - @clerk/clerk-react@5.2.7
+  - @clerk/elements@0.10.1
+
 ## 0.1.6
 
 ### Patch Changes